+++ s3pypi/core.py
@@ -5,11 +5,14 @@ from dataclasses import dataclass
 from itertools import groupby
 from operator import attrgetter
 from pathlib import Path
-from typing import List
+from typing import List, Optional
 from zipfile import ZipFile
+
+import boto3
 
 from s3pypi import __prog__
 from s3pypi.exceptions import S3PyPiError
+from s3pypi.locking import DummyLocker, DynamoDBLocker
 from s3pypi.storage import S3Storage
 
 log = logging.getLogger(__prog__)
@@ -28,27 +31,49 @@ def normalize_package_name(name: str) -> str:
     return re.sub(r"[-_.]+", "-", name.lower())
 
 
-def upload_packages(dist: List[Path], bucket: str, force: bool = False, **kwargs):
-    storage = S3Storage(bucket, **kwargs)
+def upload_packages(
+    dist: List[Path],
+    bucket: str,
+    force: bool = False,
+    lock_indexes: bool = False,
+    put_root_index: bool = False,
+    profile: Optional[str] = None,
+    region: Optional[str] = None,
+    **kwargs,
+):
+    session = boto3.Session(profile_name=profile, region_name=region)
+    storage = S3Storage(session, bucket, **kwargs)
+    lock = (
+        DynamoDBLocker(session, table=f"{bucket}-locks")
+        if lock_indexes
+        else DummyLocker()
+    )
 
     distributions = [parse_distribution(path) for path in dist]
     get_name = attrgetter("name")
 
     for name, group in groupby(sorted(distributions, key=get_name), get_name):
         directory = normalize_package_name(name)
-        index = storage.get_index(directory)
+        with lock(directory):
+            index = storage.get_index(directory)
 
-        for distr in group:
-            filename = distr.local_path.name
+            for distr in group:
+                filename = distr.local_path.name
 
-            if not force and filename in index.filenames:
-                log.warning("%s already exists! (use --force to overwrite)", filename)
-            else:
-                log.info("Uploading %s", distr.local_path)
-                storage.put_distribution(directory, distr.local_path)
-                index.filenames.add(filename)
+                if not force and filename in index.filenames:
+                    msg = "%s already exists! (use --force to overwrite)"
+                    log.warning(msg, filename)
+                else:
+                    log.info("Uploading %s", distr.local_path)
+                    storage.put_distribution(directory, distr.local_path)
+                    index.filenames.add(filename)
 
-        storage.put_index(directory, index)
+            storage.put_index(directory, index)
+
+    if put_root_index:
+        with lock(storage.root):
+            index = storage.build_root_index()
+            storage.put_index(storage.root, index)
 
 
 def parse_distribution(path: Path) -> Distribution: