+++ tests/integration/test_main.py
@@ -20,17 +20,20 @@ def test_string_dict(text, expected):
     assert string_dict(text) == expected
 
 
-def test_main_upload_package(chdir, data_dir, s3_bucket):
+def test_main_upload_package(chdir, data_dir, s3_bucket, dynamodb_table):
     with chdir(data_dir):
         dist = sorted(glob.glob("dists/*"))
-        s3pypi(*dist, "--bucket", s3_bucket.name)
+        s3pypi(*dist, "--bucket", s3_bucket.name, "--lock-indexes", "--put-root-index")
 
     def read(key: str) -> bytes:
         return s3_bucket.Object(key).get()["Body"].read()
 
+    root_index = read("index.html").decode()
+
     def assert_pkg_exists(prefix: str, filename: str):
         assert read(prefix + filename)
         assert f">{filename}</a>" in read(prefix).decode()
+        assert f">{prefix.rstrip('/')}</a>" in root_index
 
     assert_pkg_exists("foo/", "foo-0.1.0.tar.gz")
     assert_pkg_exists("hello-world/", "hello_world-0.1.0-py3-none-any.whl")